requirements.txt
certifi
chardet
charset-normalizer
googletrans
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
Pillow
pyaes
Pyrogram
PySocks
requests
rfc3986
sniffio
TgCrypto
urllib3
